Cleanup warning messages related to redundant cast
Warning messages like this will no longer happen.

   [repeat] /home/bwalker/src/netbeans/enterprise/websvc.manager/src/org/netbeans/modules/websvc/manager/ui/ReflectionHelper.java:700: warning: [cast] redundant cast to JavaParameter
   [repeat]                     JavaParameter actualParameter = (JavaParameter) parameterList.get(ii);
   [repeat]                                                     ^

+++ java/j2ee.persistence/src/org/netbeans/modules/j2ee/persistence/editor/completion/CCPaintComponent.java
@@ -225,7 +225,7 @@ public class CCPaintComponent extends JPanel {
     }
     
     protected int getWidth(String s) {
-        Integer i = (Integer)widths.get(s);
+        Integer i = widths.get(s);
         if (i != null) {
             return i.intValue();
         } else {
